backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
openai>=1.10.0
newspaper3k>=0.2.8
readability-lxml>=0.8.1
notion-client>=2.2.0
prometheus-fastapi-instrumentator>=6.1.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
